Update TLOs to work with roles.

+++ crits/certificates/views.py
@@ -7,7 +7,9 @@ from django.shortcuts import render_to_response
 from django.template import RequestContext
 
 from crits.core import form_consts
-from crits.core.user_tools import user_can_view_data
+from crits.core.user_tools import user_can_view_data, get_user_permissions
+from crits.core.data_tools import json_handler
+
 from crits.certificates.forms import UploadCertificateForm
 from crits.certificates.handlers import handle_cert_file
 from crits.certificates.handlers import delete_cert, get_certificate_details
@@ -27,6 +29,12 @@ def certificates_listing(request,option=None):
 
     if option == "csv":
         return generate_cert_csv(request)
+    elif option== "jtdelete" and not get_user_permissions(request.user.username, 'Certificate')['delete']:
+        result = {'sucess':False,
+                  'message':'User does not have permission to delete Certificate.'}
+        return HttpResponse(json.dumps(result,
+                                       default=json_handler),
+                            content_type="application/json")
     return generate_cert_jtable(request, option)
 
 @user_passes_test(user_can_view_data)
@@ -63,25 +71,30 @@ def upload_certificate(request):
     if request.method == 'POST':
         form = UploadCertificateForm(request.user, request.POST, request.FILES)
         if form.is_valid():
-            filedata = request.FILES['filedata']
-            filename = filedata.name
-            data = filedata.read() # XXX: Should be using chunks here.
-            source = form.cleaned_data.get('source_name')
-            user = request.user.username
-            description = form.cleaned_data.get('description', '')
-            related = form.cleaned_data.get('related_id', '')
-            related_type = form.cleaned_data.get('related_type', '')
-            relationship_type = form.cleaned_data.get('relationship_type','')
-            bucket_list = form.cleaned_data.get(form_consts.Common.BUCKET_LIST_VARIABLE_NAME)
-            ticket = form.cleaned_data.get(form_consts.Common.TICKET_VARIABLE_NAME)
-            method = form.cleaned_data.get('source_method', '') or 'Upload'
-            reference = form.cleaned_data.get('source_reference', '')
-            tlp = form.cleaned_data.get('source_tlp', None)
-            status = handle_cert_file(filename, data, source, user, description,
-                                      related_id=related, related_type=related_type,
-                                      relationship_type=relationship_type, method=method,
-                                      reference=reference, tlp=tlp, bucket_list=bucket_list,
-                                      ticket=ticket)
+            if get_user_permissions(request.user.username, 'Certificate')['write']:
+                filedata = request.FILES['filedata']
+                filename = filedata.name
+                data = filedata.read() # XXX: Should be using chunks here.
+                source = form.cleaned_data.get('source_name')
+                user = request.user.username
+                description = form.cleaned_data.get('description', '')
+                related = form.cleaned_data.get('related_id', '')
+                related_type = form.cleaned_data.get('related_type', '')
+                relationship_type = form.cleaned_data.get('relationship_type','')
+                bucket_list = form.cleaned_data.get(form_consts.Common.BUCKET_LIST_VARIABLE_NAME)
+                ticket = form.cleaned_data.get(form_consts.Common.TICKET_VARIABLE_NAME)
+                method = form.cleaned_data.get('source_method', '') or 'Upload'
+                reference = form.cleaned_data.get('source_reference', '')
+                tlp = form.cleaned_data.get('source_tlp', None)
+                status = handle_cert_file(filename, data, source, user, description,
+                                          related_id=related, related_type=related_type,
+                                          relationship_type=relationship_type, method=method,
+                                          reference=reference, tlp=tlp, bucket_list=bucket_list,
+                                          ticket=ticket)
+            else:
+                status = {'success':False,
+                          'message':'User does not have permission to add Certificate.'}
+
             if status['success']:
                 return render_to_response('file_upload_response.html',
                                           {'response': json.dumps({
@@ -115,8 +128,11 @@ def remove_certificate(request, md5):
     :type md5: str
     :returns: :class:`django.http.HttpResponse`
     """
-
-    result = delete_cert(md5, '%s' % request.user.username)
+    if get_user_permissions(request.user.username, 'Certificate')['delete']:
+        result = delete_cert(md5, '%s' % request.user.username)
+    else:
+        result = {'success':False,
+                  'message':'User does not have permission to delete certificate.'}
     if result:
         return HttpResponseRedirect(reverse('crits.certificates.views.certificates_listing'))
     else:
